Port `schema_settings` tests to new integration test style
