+++ schemars/tests/integration/snapshots/schemars/tests/integration/settings.rs~draft07.ser.json
@@ -1,0 +1,91 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "OuterStruct",
+  "type": "object",
+  "properties": {
+    "maybe_int": {
+      "type": [
+        "integer",
+        "null"
+      ],
+      "format": "int32",
+      "examples": [
+        8,
+        null
+      ]
+    },
+    "values": {
+      "type": "object",
+      "additionalProperties": true
+    },
+    "value": true,
+    "inner": {
+      "$ref": "#/definitions/InnerEnum"
+    },
+    "maybe_inner": {
+      "anyOf": [
+        {
+          "$ref": "#/definitions/InnerEnum"
+        },
+        {
+          "type": "null"
+        }
+      ]
+    },
+    "tuples": {
+      "type": "array",
+      "items": {
+        "type": "array",
+        "maxItems": 2,
+        "minItems": 2,
+        "items": [
+          {
+            "type": "integer",
+            "format": "uint8",
+            "minimum": 0
+          },
+          {
+            "type": "integer",
+            "format": "int64"
+          }
+        ]
+      }
+    }
+  },
+  "required": [
+    "maybe_int",
+    "values",
+    "value",
+    "inner",
+    "maybe_inner",
+    "tuples"
+  ],
+  "definitions": {
+    "InnerEnum": {
+      "oneOf": [
+        {
+          "type": "string",
+          "enum": [
+            "UndocumentedUnit1",
+            "UndocumentedUnit2"
+          ]
+        },
+        {
+          "description": "This is a documented unit variant",
+          "type": "string",
+          "const": "DocumentedUnit"
+        },
+        {
+          "type": "object",
+          "properties": {
+            "ValueNewType": true
+          },
+          "required": [
+            "ValueNewType"
+          ],
+          "additionalProperties": false
+        }
+      ]
+    }
+  }
+}
