+++ schemars/tests/integration/snapshots/schemars/tests/integration/settings.rs~draft2020_12.de.json
@@ -1,10 +1,13 @@
 {
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Outer",
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "title": "OuterStruct",
   "type": "object",
   "properties": {
-    "int": {
-      "type": "integer",
+    "maybe_int": {
+      "type": [
+        "integer",
+        "null"
+      ],
       "format": "int32",
       "examples": [
         8,
@@ -17,9 +20,12 @@
     },
     "value": true,
     "inner": {
+      "$ref": "#/$defs/InnerEnum"
+    },
+    "maybe_inner": {
       "anyOf": [
         {
-          "$ref": "#/definitions/Inner"
+          "$ref": "#/$defs/InnerEnum"
         },
         {
           "type": "null"
@@ -30,9 +36,7 @@
       "type": "array",
       "items": {
         "type": "array",
-        "maxItems": 2,
-        "minItems": 2,
-        "items": [
+        "prefixItems": [
           {
             "type": "integer",
             "format": "uint8",
@@ -42,18 +46,20 @@
             "type": "integer",
             "format": "int64"
           }
-        ]
+        ],
+        "minItems": 2,
+        "maxItems": 2
       }
     }
   },
   "required": [
-    "int",
     "values",
     "value",
+    "inner",
     "tuples"
   ],
-  "definitions": {
-    "Inner": {
+  "$defs": {
+    "InnerEnum": {
       "oneOf": [
         {
           "type": "string",
